Fix `undefined constant ConnectionBuilder` compilation error
Compiling the TDS driver with Crystal v1.14.0 results in the
following error:

    C:\crystal-tds>crystal spec
    Showing last frame. Use --error-trace for full trace.

    In src\tds\driver.cr:13:5

     13 | ConnectionBuilder.new(connection_options(params), TDS::Connection::Options.from_uri(uri))
          ^----------------
    Error: undefined constant ConnectionBuilder

The `ConnectionBuilder` class had inadvertently been declared with
an extra `TDS` level in its namespace by mistake:

    TDS::Driver::TDS::ConnectionBuilder

Then it was later referred to using `TDS::ConnectionBuilder` which
somehow worked on previous versions of the compiler.

This fix removes the extra `TDS` level from the `ConnectionBuilder`
namespace, as this was a typo, and changes references to use the
fully-qualified name `TDS::Driver::ConnectionBuilder`.