Fixes `PreparedStatement` and `UnpreparedStatement` handling of parameter placeholder `?` to correctly support queries using quoted values or comments legitimately containing `?`.
Previously the logic used a naive regular expression match against any
occurrence of `?`. The new logic uses a simple query parser which detects
and ignores quoted values and comments.

Parameter placeholders `?` are now replaced with `@PARAM_CRYSTAL_TDS_$n__`
where n = 0..n for the number of parameters specified in the query. This
parameter naming should reduce the possibility of clashes with existing
variable declarations in SQL statements.

`UnpreparedStatement` now uses `sp_executesql` to execute queries, allowing
for parameter values to be bound rather than the previous approach of inlining
the values which is a SQL injection vector.

Also adds support for `$1` .. `$n` parameter placeholders, which can now be
used instead of `?` if desired. Use of `$n` placeholders makes it possible to
reuse a parameter in multiple places in a statement. Note that mixing the use
of `?` and `$n` within a statement is not allowed and will raise an exception.

Resolves #22